+++ src/modules/plugins/plugins.gateway.ts
@@ -1,13 +1,22 @@
 import { EventEmitter } from 'events';
-import { UseGuards } from '@nestjs/common';
+import { UseGuards, UsePipes, ValidationPipe } from '@nestjs/common';
 import { SubscribeMessage, WebSocketGateway, WsException } from '@nestjs/websockets';
 import * as color from 'bash-color';
+
 import { PluginsService } from './plugins.service';
+import { PluginActionDto, HomebridgeUpdateActionDto } from './plugins.dto';
 import { Logger } from '../../core/logger/logger.service';
 import { WsAdminGuard } from '../../core/auth/guards/ws-admin-guard';
 
 @UseGuards(WsAdminGuard)
 @WebSocketGateway({ namespace: '/plugins' })
+@UsePipes(new ValidationPipe({
+  whitelist: true,
+  exceptionFactory: ((err) => {
+    console.error(err);
+    return new WsException(err);
+  })
+}))
 export class PluginsGateway {
 
   constructor(
@@ -16,9 +25,9 @@ export class PluginsGateway {
   ) { }
 
   @SubscribeMessage('install')
-  async installPlugin(client: EventEmitter, payload: string) {
+  async installPlugin(client: EventEmitter, payload: PluginActionDto) {
     try {
-      return await this.pluginsService.installPlugin(payload, client);
+      return await this.pluginsService.installPlugin(payload.name, payload.version || 'latest', client);
     } catch (e) {
       this.logger.error(e);
       client.emit('stdout', '\n\r' + color.red(e.toString()) + '\n\r');
@@ -27,9 +36,9 @@ export class PluginsGateway {
   }
 
   @SubscribeMessage('uninstall')
-  async uninstallPlugin(client: EventEmitter, payload: string) {
+  async uninstallPlugin(client: EventEmitter, payload: PluginActionDto) {
     try {
-      return await this.pluginsService.uninstallPlugin(payload, client);
+      return await this.pluginsService.uninstallPlugin(payload.name, client);
     } catch (e) {
       this.logger.error(e);
       client.emit('stdout', '\n\r' + color.red(e.toString()) + '\n\r');
@@ -38,9 +47,9 @@ export class PluginsGateway {
   }
 
   @SubscribeMessage('update')
-  async updatePlugin(client: EventEmitter, payload: string) {
+  async updatePlugin(client: EventEmitter, payload: PluginActionDto) {
     try {
-      return await this.pluginsService.updatePlugin(payload, client);
+      return await this.pluginsService.updatePlugin(payload.name, payload.version || 'latest', client);
     } catch (e) {
       this.logger.error(e);
       client.emit('stdout', '\n\r' + color.red(e.toString()) + '\n\r');
@@ -49,9 +58,9 @@ export class PluginsGateway {
   }
 
   @SubscribeMessage('homebridge-update')
-  async homebridgeUpdate(client: EventEmitter) {
+  async homebridgeUpdate(client: EventEmitter, payload: HomebridgeUpdateActionDto) {
     try {
-      return await this.pluginsService.updateHomebridgePackage(client);
+      return await this.pluginsService.updateHomebridgePackage(payload.version || 'latest', client);
     } catch (e) {
       this.logger.error(e);
       client.emit('stdout', '\n\r' + color.red(e.toString()) + '\n\r');