+++ test/e2e/backup.e2e-spec.ts
@@ -160,7 +160,7 @@ describe('BackupController (e2e)', () => {
 
     expect(client.emit).toBeCalledWith('stdout', expect.stringContaining('Restoring backup'));
     expect(client.emit).toBeCalledWith('stdout', expect.stringContaining('Restore Complete'));
-    expect(pluginsService.installPlugin).toBeCalledWith('homebridge-mock-plugin', client);
+    expect(pluginsService.installPlugin).toBeCalledWith('homebridge-mock-plugin', expect.anything(), client);
 
     // ensure the temp restore directory was removed
     expect(await fs.pathExists(restoreDirectory)).toEqual(false);